# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一个基于Go和Gin的博客系统API文档",
        "title": "Bmtdblog API",
        "termsOfService": "http://localhost:8090",
        "contact": {
            "name": "API Support",
            "url": "http://localhost:8090",
            "email": "3138910969@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8090",
    "basePath": "/",
    "paths": {
...
        "/admin/link/create": {
            "post": {
                "description": "创建新的友情链接",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "链接管理"
                ],
                "summary": "创建链接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "链接名称",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "链接URL",
                        "name": "url",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "排序值",
                        "name": "sort",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"succeed\":false,\"message\":\"错误信息\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/link/{id}": {
            "post": {
                "description": "更新已存在的友情链接信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "链接管理"
                ],
                "summary": "更新链接",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "链接ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "链接名称",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "链接URL",
                        "name": "url",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "排序值",
                        "name": "sort",
                        "in": "body",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"succeed\":false,\"message\":\"错误信息\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/link/{id}/delete": {
            "post": {
                "description": "删除指定ID的友情链接",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "友情链接管理"
                ],
                "summary": "删除友情链接",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "友情链接ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"succeed\":false,\"message\":\"错误信息\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/admin/tag/create": {
            "post": {
                "description": "创建新的文章标签",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "标签管理"
                ],
                "summary": "创建标签",
                "parameters": [
                    {
                        "type": "string",
                        "description": "标签名称",
                        "name": "value",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"succeed\":false,\"message\":\"错误信息\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/auth/{authType}": {
            "get": {
                "description": "根据认证类型获取第三方登录URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户认证"
                ],
                "summary": "获取认证URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证类型，如github",
                        "name": "authType",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "重定向到认证页面",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/oauth2callback": {
            "get": {
                "description": "处理GitHub OAuth2认证回调，完成用户登录或绑定",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户认证"
                ],
                "summary": "GitHub OAuth回调",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GitHub授权码",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "状态参数，用于防止CSRF攻击",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "301": {
                        "description": "认证失败后重定向到登录页",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/signin": {
            "get": {
                "description": "显示用户登录表单页面",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "用户认证"
                ],
                "summary": "用户登录页面",
                "responses": {
                    "200": {
                        "description": "登录页面",
                        "schema": {
                            "type": "html"
                        }
                    }
                }
            },
            "post": {
                "description": "使用用户名和密码登录系统",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "用户认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录失败，返回错误信息",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "302": {
                        "description": "登录成功，重定向到相应页面",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}